Divided tests for RestClientExtensionsTests
